refactor: newtabのロジックをhooks/サブコンポーネントに抽出
Closes #92
