+++ src/hooks/index.ts
@@ -1,7 +1,9 @@
 export * from './useAnalytics';
+export * from './useBackgroundPreload';
+export * from './useBackgroundStats';
 export * from './useBlocklist';
 export * from './useLicense';
-export * from './useSchedules';
 export * from './usePresets';
-export * from './useBackgroundStats';
 export * from './usePremiumStatus';
+export * from './useResolvedPreset';
+export * from './useSchedules';
